Upgrade jhipster to 2.8.2
We upgraded the application opt jhipster 2.8.2.
The official upgrade tool wanted to upgrade to spring-webflux,
which is not needed. We decided use an empty jhipster 2.8.2
template to port the normal spring-web components.

Angular was upgraded from 5 to 7, so many libraries changed.

This commit will need further testing, since we maybe have
overwritten important changes.

Author: NiHa, LuGe